requirements.txt
groq
langchain_groq
langgraph
langchain-core
langsmith

PyPDF2
requests

python-dotenv
typing-extensions

streamlit